binary map being passed

+++ src/rwnmr/framework/CustomMatrix.h
@@ -9,21 +9,21 @@ public:
     CustomMat(int rows, int cols) : rows(rows), cols(cols) {
         data.resize(rows * cols, 0.0);
     }
-    CustomMat(int rows, int cols, const std::vector<double>& data) : rows(rows), cols(cols), data(data) {}
+    CustomMat(int rows, int cols, const std::vector<uint8_t>& data) : rows(rows), cols(cols), data(data) {}
 
-    double& at(int row, int col) {
+    uint8_t& at(int row, int col) {
         return data[row * cols + col];
     }
 
-    const double& at(int row, int col) const {
+    const uint8_t& at(int row, int col) const {
         return data[row * cols + col];
     }
-    static std::vector<double> ones(int rows, int cols) {
-        std::vector<double> ones(rows * cols, 1.0);
+    static std::vector<uint8_t> ones(int rows, int cols) {
+        std::vector<uint8_t> ones(rows * cols, 1.0);
         return ones;
     }
-    static std::vector<double> eye(int rows, int cols) {
-        std::vector<double> eye(rows * cols, 0.0);
+    static std::vector<uint8_t> eye(int rows, int cols) {
+        std::vector<uint8_t> eye(rows * cols, 0.0);
         for (int i = 0; i < std::min(rows, cols); i++) {
             eye[i * cols + i] = 1.0;
         }
@@ -32,13 +32,13 @@ public:
 
     int getRows() const { return rows; }
     int getCols() const { return cols; }
-    const std::vector<double>& getData() const {
+    const std::vector<uint8_t>& getData() const {
         return data;
     }
 
 private:
     int rows, cols;
-    std::vector<double> data;
+    std::vector<uint8_t> data;
 };
 
 #endif
